Update package.js
Update to Vulcan. The current version of this package is actually dependant on `ui-boostrap`, because this package handle the first registeration of the component with `registerComponent`. Without it, `replaceComponent` will fail.

However, this will be fixed in Vulcan 1.11.3, which tolerate a call to `replaceComponent` even if the component does not exist yet.